Fixed bugs!
I fixed the seg fault you were having by changing the way you hand
shuffled the cards (your code is still there but currently commented
out). Instead of using a linked list I just created a copy of the deck
array and manipulated it. It was throwing an error on the sort
functions because your way produced duplicates of some cards and
skipped others.
I also made it so that the menu printed after every option, since it
was difficult to remember which option did what. I also fixed your exit
option because it wasn’t exiting the program when you selected 8.
Finally, the way you were using the rand function made it so that the
program shuffled 5 times all of the time, since rand()% 1 + 5 will
always be 5. I changed it to rand()%10 + 5, so it will shuffle between
5 and 15 times.
Hope this helps! Please let me know if anything I did breaks your code,
but I think it’s all working now. My email is paar5780@colorado.edu.
~Paige